fileName: IndexRoute.kt, EmployeeService.kt reason: alter the type of "id" from Long to Int.

+++ src/main/kotlin/com/xust/plugins/route/IndexRoute.kt
@@ -61,7 +61,7 @@ fun Route.home() {
 fun Route.getEmployeeById() {
     get("/employee/{id}") {
         val id = call.parameters["id"]
-        val employeeList = getEmployeeByIdService(id!!.toLong())
+        val employeeList = getEmployeeByIdService(id!!.toInt())
         call.respondText(employeeList[0].toString())
     }
 }
